django-rest-framework/requirements.txt
beautifulsoup4==4.5.3
requests==2.22.0
lxml==4.3.0
Django==2.2.4
django-cors-headers==3.1.0
djangorestframework==3.10.2
requests-cache==0.5.2